Added caching support In an effort to make Iridium the go-to ORM for Node.js applications :wink: we've implemented inline caching support - allowing you to easily cache requests for single objects using your favourite caching engine (Redis/Memcache etc.).
As it stands, the cache is only hit for Model.get(_id)/Model.find(_id) and is automatically
updated when inserts and instance specific saves/refreshes/removals are executed.

It doesn't aim to be a be-all and end-all replacement for having a fast MongoDB server,
but it should allow you to offload some of the overhead on common operations to some
kind of KVS which is better suited to the task.