Initial push of TypeScript rewrite
This rewrite is now feature complete - on part with v4.0.0 - however it
has not yet been tested AT ALL. IntelliSense works (take a look at
example\IntelliSense.ts) but we still need to port all the tests over
from v4.x to vNext and document the process for upgrading existing
applications (both TypeScript and JavaScript). We then need to tweak the
README file to document the new API and give new usage examples.