Removed transforms framework
This is a big change, but it's been done for a couple of reasons. The
first is that it helps make Iridium quite a bit faster - not doing up
and down transformations everywhere, and helps reduce the likelihood of
V8 de-optimizing code because of continuously changing objects. It also
means that we can simplify the pipeline and in most cases it is still
possible to achieve the same transforms using getters/setters (which are
now easier to deal with) and the instance.document object (which is now
always in the same format as the database version).