refactor: Update to use TypeScript 2.0 :smile:  (breaking)
It's finally here! :fireworks:
This refactor updates the way we perform module resolution to make use
of @types/<lib>, removing the dependency on typings, improves the way
the library is bundled while still providing type hints (smaller
downloads and faster npm installs) and makes creating validation plugins
far easier (with the addition of the this:<type> operator in TypeScript
2.0). We've also gone through the codebase and updated it to make use of
strict null checking, which should help prevent a couple of small edge
cases going forward; things that may have been missed during testing. To
compliment that, we've also added support for type checking of your
condition and update arguments, making those more intuitive to use and
more robust against typos.

All this goodness comes at the cost of possible breakage in the way you
use the API, specifically if you add support for strict null checking to
your project, which will enforce some more stringent rules regarding
when things can be used. If you're checking return values for `null` you
will also find that a number of them have been replaced with
`undefined`, hopefully this doesn't affect too many of you (because of
how easy it is to write `if (x) { ... }` instead, but be aware of the
change if you routinely write `if (x === null) { ... }` in your
handlers.