feat #118: Add intelligent memoization of transformed fields on Instances
This makes it possible to make modifications to objects which are the result of field transforms in a safe manner which is reflected in the changes made to the database. This should enable complex objects to work intuitively through transforms without the need for re-assigning to the field when you are done.

Additionally, this fixes an issue with transformed fields when an instance is saved for the first time - by both fixing the bug that caused it and strictly defining the roles of TDocument and TInstance (with TDocument now representing the database-types and TInstance representing the post-transform types).